Activities grouping happened now.
