+++ src/main/java/com/gaadikey/gaadikey/gaadikey/Activities/BaseSampleActivity.java
@@ -1,10 +1,12 @@
-package com.gaadikey.gaadikey.gaadikey;
+package com.gaadikey.gaadikey.gaadikey.Activities;
 
 import android.support.v4.app.FragmentActivity;
 import android.support.v4.view.ViewPager;
 import android.view.MenuItem;
 import android.widget.Toast;
 
+import com.gaadikey.gaadikey.gaadikey.R;
+import com.gaadikey.gaadikey.gaadikey.TestFragmentAdapter;
 import com.viewpagerindicator.PageIndicator;
 
 import java.util.Random;
