+++ src/main/java/com/gaadikey/gaadikey/gaadikey/Activities/FriendsLaneActivity.java
@@ -1,4 +1,4 @@
-package com.gaadikey.gaadikey.gaadikey;
+package com.gaadikey.gaadikey.gaadikey.Activities;
 
 import android.app.ListActivity;
 import android.content.SharedPreferences;
@@ -11,6 +11,8 @@ import android.view.View;
 import android.widget.ListView;
 import android.widget.Toast;
 
+import com.gaadikey.gaadikey.gaadikey.R;
+import com.gaadikey.gaadikey.gaadikey.ViewNotifyObject;
 import com.gaadikey.gaadikey.gaadikey.adaptor.FriendsAdapter;
 
 import org.apache.http.HttpResponse;
