+++ src/main/java/com/gaadikey/gaadikey/gaadikey/Activities/IntroActivity.java
@@ -1,11 +1,14 @@
-package com.gaadikey.gaadikey.gaadikey;
+package com.gaadikey.gaadikey.gaadikey.Activities;
 
 import android.os.Bundle;
 import android.support.v4.view.ViewPager;
 
+import com.gaadikey.gaadikey.gaadikey.Activities.BaseSampleActivity;
+import com.gaadikey.gaadikey.gaadikey.GaadiKey;
+import com.gaadikey.gaadikey.gaadikey.R;
+import com.gaadikey.gaadikey.gaadikey.TestFragmentAdapter;
 import com.google.android.gms.analytics.HitBuilders;
 import com.google.android.gms.analytics.Tracker;
-import com.viewpagerindicator.CirclePageIndicator;
 
 public class IntroActivity extends BaseSampleActivity {
