+++ src/main/java/com/gaadikey/gaadikey/gaadikey/Activities/MyActivity.java
@@ -1,4 +1,4 @@
-package com.gaadikey.gaadikey.gaadikey;
+package com.gaadikey.gaadikey.gaadikey.Activities;
 
 import android.app.Activity;
 import android.app.AlertDialog;
@@ -28,6 +28,10 @@ import android.widget.Spinner;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.gaadikey.gaadikey.gaadikey.Constants;
+import com.gaadikey.gaadikey.gaadikey.GaadiKey;
+import com.gaadikey.gaadikey.gaadikey.ProfileObject;
+import com.gaadikey.gaadikey.gaadikey.R;
 import com.google.android.gms.analytics.HitBuilders;
 import com.google.android.gms.analytics.Tracker;
 import com.google.android.gms.gcm.GoogleCloudMessaging;
