+++ src/main/java/com/gaadikey/gaadikey/gaadikey/Activities/RadioActivity.java
@@ -1,26 +1,34 @@
-package com.gaadikey.gaadikey.gaadikey;
+package com.gaadikey.gaadikey.gaadikey.Activities;
 
-import android.support.v7.app.ActionBarActivity;
 import android.os.Bundle;
+import android.support.v7.app.ActionBarActivity;
+import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 
+import com.gaadikey.gaadikey.gaadikey.R;
 
-public class ContactsActivity extends ActionBarActivity {
+
+public class RadioActivity extends ActionBarActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_contacts);
+        setContentView(R.layout.activity_radio);
+      //  AlertDialogView();
+        Log.e("We are at Radio Activity" , "The Radio Activity is here "); // consoling that it has entered radio activity!
+
     }
 
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.contacts, menu);
+        getMenuInflater().inflate(R.menu.radio, menu);
         return true;
     }
+
+
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
