+++ src/main/java/com/gaadikey/gaadikey/gaadikey/Activities/SettingUpActivity.java
@@ -1,4 +1,4 @@
-package com.gaadikey.gaadikey.gaadikey;
+package com.gaadikey.gaadikey.gaadikey.Activities;
 
 import android.content.Intent;
 import android.content.SharedPreferences;
@@ -12,6 +12,10 @@ import android.view.Menu;
 import android.view.MenuItem;
 import android.widget.Toast;
 
+import com.gaadikey.gaadikey.gaadikey.ContactsObject;
+import com.gaadikey.gaadikey.gaadikey.GaadiKey;
+import com.gaadikey.gaadikey.gaadikey.LaunchActivity_NavDrawer;
+import com.gaadikey.gaadikey.gaadikey.R;
 import com.google.android.gms.analytics.HitBuilders;
 import com.google.android.gms.analytics.Tracker;
