+++ src/main/java/com/gaadikey/gaadikey/gaadikey/GaadiKey.java
@@ -31,7 +31,7 @@ public class GaadiKey extends Application {
         super();
     }
 
-    synchronized Tracker getTracker(TrackerName trackerId) {
+    public synchronized Tracker getTracker(TrackerName trackerId) {
         if (!mTrackers.containsKey(trackerId)) {
 
             GoogleAnalytics analytics = GoogleAnalytics.getInstance(this);
