+++ src/main/java/com/gaadikey/gaadikey/gaadikey/TestFragmentAdapter.java
@@ -8,7 +8,7 @@ import android.util.Log;
 import com.viewpagerindicator.IconPagerAdapter;
 
 
-class TestFragmentAdapter extends FragmentPagerAdapter implements IconPagerAdapter {
+public class TestFragmentAdapter extends FragmentPagerAdapter implements IconPagerAdapter {
     protected static final String[] CONTENT = new String[] { "intro1", "intro2", "intro3", "intro4" };
     protected static final int[] ICONS = new int[] {
             R.drawable.perm_group_calendar,
