+++ src/main/res/layout/activity_contacts.xml
@@ -6,7 +6,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="com.gaadikey.gaadikey.gaadikey.ContactsActivity">
+    tools:context=".Activities.ContactsActivity">
 
     <TextView
         android:text="@string/hello_world"
