+++ src/main/res/layout/activity_enter_pin.xml
@@ -6,7 +6,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="com.gaadikey.gaadikey.gaadikey.EnterPINActivity">
+    tools:context=".Activities.EnterPINActivity">
 
     <TextView
         android:text="Enter the PIN which you received by SMS/Email. (You will receive the SMS within 5 minutes.) "
