+++ src/main/res/layout/activity_friends_lane.xml
@@ -6,7 +6,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="com.gaadikey.gaadikey.gaadikey.FriendsLaneActivity">
+    tools:context=".Activities.FriendsLaneActivity">
 
     <TextView
         android:id="@+id/tipBox"
