+++ src/main/res/layout/activity_intro.xml
@@ -20,7 +20,7 @@ xmlns:android="http://schemas.android.com/apk/res/android"
 android:orientation="vertical"
 android:layout_width="fill_parent"
 android:layout_height="fill_parent"
-tools:context="com.gaadikey.gaadikey.gaadikey.IntroActivity">
+tools:context=".Activities.IntroActivity">
 
 <android.support.v4.view.ViewPager
     android:id="@+id/pager"
