+++ src/main/res/layout/activity_setting_up.xml
@@ -6,7 +6,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="com.gaadikey.gaadikey.gaadikey.SettingUpActivity">
+    tools:context=".Activities.SettingUpActivity">
 
     <TextView
         android:text="Setting up your account! \n Please wait while we setup your account."
