+++ src/main/res/layout/activity_shoppinglane.xml
@@ -6,7 +6,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="com.gaadikey.gaadikey.gaadikey.ShoppinglaneActivity">
+    tools:context=".Activities.ShoppinglaneActivity">
     <TextView
 
         android:layout_width="wrap_content"
