+++ src/main/res/layout/activity_verification.xml
@@ -6,7 +6,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="com.gaadikey.gaadikey.gaadikey.VerificationActivity">
+    tools:context=".Activities.VerificationActivity">
     <TextView
         android:text="In order to proceed we have to verify your phone number. Once the phone number is verified you can build your GaadiKey profile. "
         android:layout_width="wrap_content"
