+++ src/main/res/layout/fragment_base_sample.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.gaadikey.gaadikey.gaadikey.BaseSampleActivity">
+    tools:context=".Activities.BaseSampleActivity">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
