+++ src/main/res/xml/ecommerce_tracker.xml
@@ -7,5 +7,5 @@
     <!-- Treat events as test events and don't send to google -->
     <bool name="ga_dryRun">false</bool>
     <!-- The screen names that will appear in reports -->
-    <string name="com.gaadikey.gaadikey.gaadikey.MyActivity">ActivityLauncher</string>
+    <string name="com.gaadikey.gaadikey.gaadikey.Activities.MyActivity">ActivityLauncher</string>
 </resources>
